+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/programrule/RuleActionEventMapper.java
@@ -53,15 +53,15 @@ import org.hisp.dhis.setting.SystemSettingManager;
 import org.hisp.dhis.tracker.bundle.TrackerBundle;
 import org.hisp.dhis.tracker.domain.DataValue;
 import org.hisp.dhis.tracker.domain.Event;
-import org.hisp.dhis.tracker.programrule.implementers.RuleActionExecutor;
-import org.hisp.dhis.tracker.programrule.implementers.ValidationRuleAction;
-import org.hisp.dhis.tracker.programrule.implementers.event.AssignDataValueExecutor;
-import org.hisp.dhis.tracker.programrule.implementers.event.RuleEngineErrorExecutor;
-import org.hisp.dhis.tracker.programrule.implementers.event.SetMandatoryFieldExecutor;
-import org.hisp.dhis.tracker.programrule.implementers.event.ShowErrorExecutor;
-import org.hisp.dhis.tracker.programrule.implementers.event.ShowErrorOnCompleteExecutor;
-import org.hisp.dhis.tracker.programrule.implementers.event.ShowWarningExecutor;
-import org.hisp.dhis.tracker.programrule.implementers.event.ShowWarningOnCompleteExecutor;
+import org.hisp.dhis.tracker.programrule.executor.RuleActionExecutor;
+import org.hisp.dhis.tracker.programrule.executor.ValidationRuleAction;
+import org.hisp.dhis.tracker.programrule.executor.event.AssignDataValueExecutor;
+import org.hisp.dhis.tracker.programrule.executor.event.RuleEngineErrorExecutor;
+import org.hisp.dhis.tracker.programrule.executor.event.SetMandatoryFieldExecutor;
+import org.hisp.dhis.tracker.programrule.executor.event.ShowErrorExecutor;
+import org.hisp.dhis.tracker.programrule.executor.event.ShowErrorOnCompleteExecutor;
+import org.hisp.dhis.tracker.programrule.executor.event.ShowWarningExecutor;
+import org.hisp.dhis.tracker.programrule.executor.event.ShowWarningOnCompleteExecutor;
 import org.springframework.stereotype.Service;
 
 @Service( "org.hisp.dhis.tracker.programrule.RuleActionEventMapper" )
