--- dhis-2/dhis-services/dhis-service-tracker/src/test/java/org/hisp/dhis/tracker/programrule/implementers/event/RuleEngineErrorExecutorTest.java
@@ -1,81 +1,0 @@
-/*
- * Copyright (c) 2004-2022, University of Oslo
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- * Neither the name of the HISP project nor the names of its contributors may
- * be used to endorse or promote products derived from this software without
- * specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package org.hisp.dhis.tracker.programrule.implementers.event;
-
-import static org.hisp.dhis.tracker.programrule.ProgramRuleIssue.warning;
-import static org.hisp.dhis.tracker.validation.ValidationCode.E1300;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import java.util.Optional;
-
-import org.hisp.dhis.DhisConvenienceTest;
-import org.hisp.dhis.tracker.bundle.TrackerBundle;
-import org.hisp.dhis.tracker.domain.Event;
-import org.hisp.dhis.tracker.programrule.ProgramRuleIssue;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-@ExtendWith( MockitoExtension.class )
-class RuleEngineErrorExecutorTest extends DhisConvenienceTest
-{
-    private final static String RULE_ID = "Rule_id";
-
-    private final static String ERROR_MESSAGE = "Error message";
-
-    private final static String EVENT_ID = "EventUid";
-
-    private final RuleEngineErrorExecutor ruleEngineErrorExecutor = new RuleEngineErrorExecutor( RULE_ID,
-        ERROR_MESSAGE );
-
-    private TrackerBundle bundle;
-
-    @BeforeEach
-    void setUpTest()
-    {
-        bundle = TrackerBundle.builder().build();
-    }
-
-    @Test
-    void shouldReturnAWarningWhenRuleWithSyntaxErrorIsTriggeredOnEvent()
-    {
-        Optional<ProgramRuleIssue> warning = ruleEngineErrorExecutor.executeRuleAction( bundle, getEvent() );
-
-        assertTrue( warning.isPresent() );
-        assertEquals( warning( RULE_ID, E1300, ERROR_MESSAGE ), warning.get() );
-    }
-
-    private Event getEvent()
-    {
-        return Event.builder()
-            .event( EVENT_ID )
-            .build();
-    }
-}
